Component for notification bubbles
Notifications bubbles now use a NotificationWindow component which
is a QQuickWindow subclass that sets the surface format and has a
public slot that sends the wl_notification_daemon_add_surface
request to the compositor.

Here's how notifications work:

NotificationDaemon implements the XDG specification, it receives
a request to show a notfication bubble via DBus and checks if
the notification was already created (in this case it updates
the body). If it's not, it creates a NotificationItem and shows
it.

After a notification bubble is shown, we process the events to let
the wayland QPA plugin create the buffer and attach the surface.
After that we send the wl_notification_daemon_add_surface request
to the compositor which properly set the position.

Issue: #6