Image provider for desktop entries
Custom image provider for application icons.
If a desktop entry icon is not found in the current theme, we verify if
it's an absolute path and in that case we load the pixmap.
Otherwise we assume it's DATADIR/name.png, if this path doesn't exist
or the pixmap is invalid we fallback to application-x-executable.

Issue: #21