Refactor Launcher window
Due to single QML engine changes we create shell windows from C++, because
it's easier as we have one single place (ShellUi) where we have the engine
and the screen.

While we are here, We fix multi screen issues by not hardcoding X and Y
coordinates to 0, but using actual monitor X and Y coordinates.

Issue: #22
Issue: #23