Always launch a new instance of pinned applications
Since we don't have a proper task management we don't know when an
application is closed. If we had better task management we would be able
to activate running applications and launch new instances for closed
applications. We workaround the lack of task management launching new
instances all the time.

Issue: #25