Don't hide lock screen before closing it
Hiding the lock screen surface seems to have problems: the surface get
stuck until the mouse is moved and users would think the session is
still locked.

This might be related to a qtwayland problem, see:
  https://codereview.qt-project.org/#change,59262

Until the situation becomes more clear we don't hide the surface, it
will be destroyed by deleteLater() eventually.