Trigger AppChooser with Ctrl+Esc
Meta key is already used as a modifier for shell actions and Weston
can't reuse it for another binding.

Being a single key press it might be possible to determine that Meta is
not used as a modifier, maybe we can patch Weston to do so but not
in Hawaii 0.2.0 time frame so we use Ctrl+Esc for now.

Closes #63