Fix splash surface with multiple outputs from Xorg
When running weston --output-count=2 (or any number greater than 1) from
X11 results in a black opaque surface that covers the whole desktop.

Weston already has a layer for the fade surface, let's use this instead
of using our own. This also has the side effect that the cursor is no
longer shown on top of the fade surfaces.

We also create one giant surface that covers all the outputs and
destroy it as soon as it's not needed.

This change also make createBlackSurface() public, so that Splash can
have access to it.

Closes #73