Notifications image provider
Cache images for notifications and retrieve them with an image provider
instead of using an ImageItem because we have troubles setting an image
from C++.

This feels much more clean and that's how you are supposed to use images
from the QtQuick world.

There's also the nice side effect that don't create a separate QML
engine anymore.

Closes #74