Import libqtxdg fork
This library is not usable as a dependency here because its CMake
configuration doesn't provide library path.  A pull request from my fork
was sent upstream and will hopefully be accepted soon.

Also, libqtxdg uses libmagic for MIME types and implements the base
directory specification itself instead of using new Qt 5 stuff.
It's my intention to make it take advantage of Qt 5 in my fork and send
the code upstream.  Meanwhile the library is used here as a submodule.