server: Disable shell dialog animations for now
We have problems with authentication dialog. The surface is supposed
to progressively resize to full size but this doesn't happen.

Here's what happens:

 - As authentication is requested a dialog is created.
 - The modal overlay fades in and we see a really small rectangle
   centered on the output, that should be the authentication dialog.
 - Hitting tab will give password field focus and this updates the
   buffer resulting in the dialog being resized to full size.

Expected result: dialog resizes to full size progressively.

Works with the shutdown dialog though...

I decided to disable the animation for now so that we can finally
release Hawaii 0.2!