client: Refactor Wayland integration
Move wl_hawaii_shell implementation into ShellClient and listen to
wl_registry events with RegistryListener.

HawaiiShell is renamed to ShellController and it's no longer a
singleton, it's instantiated by ShellManager which sets up the QML
engine and register types.

When the compositor wants the user interface to be created,
ShellController calls a slot in ShellManager which in turn creates
ShellUi and emit the ready() signal telling the compositor that the
shell is ready.

Issue: #122