server: Improve shell surfaces placing
With latest changes we don't call set_position() anymore for a number of
shell surfaces because it's redundant.

Background surfaces are placed according to the output coordinates and
so are Overlays. Panels on the other hand, still need improvements that
will be made later, their position is not currently set.

Issue: #124