client: Set panel geometry
When the screen geometry changes, the panel is moved to another screen
or properties are changed, recalculate the geometry and ask the
compositor to move the surface to the new position.

While we're here we add new properties: form factor and thickness (panel
size, that depends on form factor).

Issue: #122