libhawaiishell: Set incubation mode to AsynchronousIfNested by default
If the component is being created in a context that is already part of
an asynchronous creation, the incubator will join that existing
incubation and execute asynchronously.

As a result, the existing incubation won't be Ready until both have been
completed

Otherwise, if the component is being create in a syncrhonous context the
incubation will execute synchronously.

Issue: #130