desktop: Move configuration button back to the panel
Having a tool box plugin was a mistake, the configuration window was
supposed to be created from these plugins but since it need to know
view specific information it's now created by the panel or desktop
itself.

This leaves the tool box package to contain just a button that we now
move to the shell implementation just like it was meant to be at the
beginning.

Another valid reason to have a tool box package could have been
support for different platforms (desktop, tablet, ...) and that can
be even easier if the button is part of the shell package.

Issue: #143