Convert all metadata to JSON
JSON files are easier to handle because we don't need extra code from
libqtxdg to handle them (Qt 5 already provides an API).

The only reason I was reluctant and started with desktop entries was that
Transifex supports them, but we can synthesize a desktop file from
the JSON metadata with only Name and Comment key and give it to
the translation system. Such a script would also add the translations
to the JSON files as well.

Issue: #144