compositor: Layers
Add a bunch of items to serve as layers where we put surfaces.

One of the things I like most of Qt based compositor is that you can
easily implement layers with items with the proper z-index value.

Surface items just need to be created as children of the layer items
to get a propert stacking order.

Issue: #148