hawaii.xml: Overlays are not shell surfaces anymore
Move it from wl_hawaii_shell_surface to wl_hawaii_shell because
overlays are going to be part of the design language of the shell
as we'll only have one thus it will be part of the main UI not
something that other applications will create.

The output argument is removed because the overlay is going to be
shown in the current active screen.

Issue: #148