Redo workspaces linear view
StackView had a problem of disappearing workspaces, can't use a ListView
either because delegates will be destroyed as well.

The best solution is to put an Item with a repeater that aligns
workspaces like a strip and play with contentX to switch between
them.

Downside is that switch from the last workspace to the first
goes from right to left showing all workspaces in the middle instead
of going to the right as expected.