Rearrange the way shell windows are managed
Branch the code when mapping and unmapping surfaces for surfaces
belonging to applications or shell based on the fact that shell surfaces
always have only one view which reports the role.

Set z-index to layers to make sure they are stacked in the right order.

Remove the old surface roles from Compositor.

Shell surfaces are mapped only when a role is defined.