session: Map process with desktop file name
XdgDesktopFile pointers are retrieved from cache and might not be
available anymore by the time we access them.

Mapping desktop file names to process is easier.

The finished() signal was connected before the entry was added to the
map, map earlier so that it will be found when a process finish right
after it is started.

Also use deleteLater() on processes instead of delete.