session: logind backend
Add a login manager subsystem with a logind backend, instead of
using qlogind.

qlogind is not going to be a proper library any time soon plus it
has stuff we don't need.

Removing this dependency will result in a less fat hawaii-session
executable.

With this new subsystem we can easily make unit tests with the
logind backend.

This is also a chance to improve the code: we now inhibit sleep
in order to properly activate the lock screen.