session: Receive messages from the compositor
The compositor knows when the session is idle or not because it has
access to the input device, therefore it can determine whether
there is activity or not.

The session manager now creates a server socket and pass it to the
compositor which will send messages to inform about the status.

For now only SetIdle is implemented.

Previously this was done from QML using the session interface which
in turn would call the session manager via DBus.
By removing this mechanism we don't expose those details to the
rest of the world either.