Use a custom TextField without the edit menu
A widgets based edit menu will likely crash the application since it requires
an underlying windowing system but we are running in a compositor which is the
windowing system.

In the future we might want to replace the menu component with a QML based
popup menu, for now we set it to null to avoid crashes.