shell: Simplify workspace effects
It's too soon to be worried about custom effects that can be loaded by
the user. We just want two effects now: present windows and reveal
desktop.

Reimplement them in the workspace component.

Please note that effects are not complete, the chrome component on
windows is not created anymore and will be completed later.

Issue: #195