Move QML code into resources
This is more secure because nobody can change a single QML file to
alter the Hawaii behavior.

Thus it's still possible to do so but it requires rebuilding the
executable.

This also make deployment more reliable for people who builds the source,
in fact from time to time we move files around and installing over a
previous local build won't get rid of files in old locations.

Closes #200