compositor: Improve startup and shutdown sequence
Startup sequence is triggered with an event so that the D-Bus service
and objects are register when an application is already created.

Shutdown sequence is invoked when the application is about to quit
(either with a UNIX signal such as SIGTERM or SIGINT or on logout).
The sequence ensure that all applications are closed and all services
are destroyed before bringing the compositor down.