shell: Set maximum and minimum output window size and random position
When running Hawaii on X11 we want the windows position to be determined
by the window manager and the size to be fixed to 1024x768 to easily
move and manage them.  Of course windows will be maximizable to the
actual output size if the user desire.

We also want the window border, otherwise we cannot move the windows
around when Hawaii is running windowed inside another system.

This has no effect when using the greenisland QPA plugin, in fact it
will show each window into their native screen fully maximized.