Fix indicators breaking when changing alignment
Do not use Loader as it doesn't resize to fit its component.

Loading each indicator asynchronously was used to prevent the full
shell from loading if only one of the indicators failed but if that
happens we can catch the error and show the error screen instead.

Another reason was to be able to disable some indicators but we
don't have settings for that yet and we can find a better way to
do it.