Increase security preventing ptrace
Disable ptrace except for gdb.

This will prevent other processes running as the same user to attach to
hawaii and inject arbitrary code like a key logger, but allows gdb to
do so to keep debugging possible.

Higher privileged users will still be able to use ptrace but that is
not an issue related to key logger concerns because such users can
access the input devices directly anyway.

See also this for more information on ptrace:

https://wiki.ubuntu.com/SecurityTeam/Roadmap/KernelHardening#ptrace_Protection