Rework idle inhibit
Add a timer whose interval is determined by org.hawaiios.session
settings schema and an idle inibit counter on Compositor and another
one on Output.

Idle can be inhibited by applications.

When the timer is triggered and idle is not inhibited the output
is powered off with a fade to black animation. Any keyboard input or
mouse movement powers the output back on.

Keyboard input and mouse movement also restarts the idle timer in order
to avoid the output going dark when the user is doing something.
In the future we might want to do the same for mouse clicks, mouse
wheel and touch input.

Issue: #214