Improve virtual keyboard position and size for desktops
Qt Virtual Keyboard scales height based on width automatically, if we
fit the width it will also eat half the screen in height making the
text fields invisible.

Align the keyboard centered to the bottom with a width set to a quarter
of the screen width and a minimum of 768dp.

Also, make sure the password input field is not covered by the
virtual keyboard by moving it up.

Issue: #219