desktop: New QML module with parts of the shell
- Common login greeter and user delegate
- Indicator and PanelItem
- DateTimeIndicator
- Background

Move the indicators code here since they are going to
be shown on both the shell and SDDM we need them in a
library.

The standard background component can be used for wallpaper,
lock screen and login screen background.
It accepts string properties from settings that are internally
converted to the appropriate data type.