lock: Disable input events on all surfaces when locked
Disable all kind of input events included but not limited to mouse
events on all surfaces when the session is locked.

We can be sure now that keyboard, mouse and touch input events will
not leak through windows when the session is locked.

Between the mouse input catcher and the focus being always on the password
field we were reasonably sure this didn't happen, but actually disabling
the input events on surfaces is guaranteed to work.

The old mouse input catcher is still there, though it doesn't need hover
events and onPressed will just give focus to the password field so
it will never lose it.