Hide window system cursor and reparent idle dimmer
ApplicationWindow has an overlay on top of the scene and it catches
mouse events instead of WaylandMouseTracker.

To achieve the same result as a WaylandMouseTracker attached to fill
the parent we filter window mouse move events to track cursor position
and set/unset the QQuickOverlay cursor in order to enable/disable
the window system cursor.

Reparent WaylandCursorItem so that stays above the overlay.

We also reparent idle dimmer so that it covers the whole scene,
overlay included.

Closes: #19