compositor: XWayland improvements
Create a special chrome for XWayland windows that for now is a copy
of the one for Wayland windows but using the appropriate shell surface item.

Also do not use a Loader to load XWayland since it has an enabled
property that would allow us to decide whether we want it or not.