compositor: Solid color for splash screen
Splash screen is white until the background component is loaded asynchronously
and takes more time to load compared to the desktop itself, resulting in a
white rectangle that is faded into the desktop.

Use the same solid color from the Plymouth and SDDM themes instead, this will
also make the transition feel smooth.

Closes: #74