New present effect
Merge the original Papyros code, tweak and adapt to Liri internal API, do
multi screen fixes and overall improvements including centering the grid.

Instead of showing a side panel with a preview of the workspaces we now add
margins to the desktop that will be more useful later on, when we'll have
more than one workspace.

Also fixed a bug that allowed the shell to catch input even when it was
hidden in present mode.

While we are here we also comment out multiple workspaces that were
excluded anyway, due to input problems that we still need to figure out.

Closes: #20