Avoid infinite loop when moving XWayland windows
XWayland clients can suddenly send a configure notify event, we handle it
by moving the windows but changing moveItem coordinates will configure
the window with the new location, which will result in another configure
notify event.

In order to avoid an infinite loop we create xwaylandMoveItem and change
coordinates on that when dragging a window by the decoration.  This will
generate a configure notify that will result in moveItem moving.