Launcher overhaul
Turn ApplicationManager into a model with a complete list of installed
apps that can also track those that are running or pinned.

Avoid sharing apps list with different models and complex signals
connection, and likely several crashes.

Populate the models with QtConcurrent and when the menu changes, just
remove apps that are no longer in the menu instead of resetting
the model.

Closes: #36
Closes: #105