Install liri-session as is
INSTALL_BINDIR was expanded to a path that do not reference
qbs.installRoot anymore.

This is not wrong since the documentation says code should not reference
qbs.installRoot because that is the package installation root.
However in source builds the lack of qbs.installRoot in INSTALL_BINDIR
makes finding the liri-shell executable impossible.

Since the PATH environment variable should contain the location of
liri-shell in both source and package builds, we no longer run
the executable with its full path.