Multimedia keys
With the modularization effort going on, we moved PulseAudio support into
its own repository.

We no longer can access the VolumeControl from the shell and the
indicator cannot install global shortcuts.

Create a singleton object that emits signals when multimedia keys are
pressed and make it available in the root context to the shell and
indicators.

Closes: #134