OSD interface
Add a singleton that indicators can use to show OSD messages,
instead of messing with the overlays itself.

The new OnScreenDisplay singleton is used by both the shell and the
indicators because they are in the same process. If, in the future,
we move indicators out-of-process we can reuse the API and add
D-Bus communication.

Indicators calls OnScreenDisplay methods while the shell catches
the signals and update the overlay accordingly.

Closes: #133