Disallow running as root
The shell is not designed to run as root, so there is a great risk
that applications could attack it through Wayland protocol requests,
or by other means, causing stack or heap overflows and execute
random code.

Also, Wayland compositors are designed to run as unprivileged users
this is the reason why logind is often used to get access to
devices which would otherwise require root.

[ChangeLog][Shell] Disallow running as root.

Closes: #140