Screencast D-Bus interface
Reimplement the screencast system as a D-Bus interface.

Create a PipeWire stream that can be saved into a file
using GStreamer.

Screen is recorded only with our QPA plugin which implement
functions to enable or disable recording.  When recording is
enabled for a screen, an event with a QImage capture of
the screen is posted before swap buffers.

PipeWire support can be disabled, in that case the D-Bus
interface is dummy and do nothing.

Closes: #142