Try to autodetect the platform plugin
Inspect the environment to guess which platform plugin is best suited.

If we are using Qt 5.11.0 or better we take advantage of the fallback
mechanism and try wayland first, then xcb unless XDG_SESSION_TYPE is
empty. In that case we assume we are running under vt, and use liri.

For Qt before 5.11.0 we use wayland if WAYLAND_DISPLAY is set,
otherwise xcb if DISPLAY is set, or fallback to liri.

Closes: #150