Don't destroy wl-shell surfaces for closing
wl-shell doesn't have a specific request to close windows, destroying
the surface causes the compositor to crash.

As soon as we can use xdg-shell this problem will go away for the
vast majority of clients, hence it's safer to warn users that close
is not supported and avoid crashing the server.