Window management overhaul due to xdg-shell-v6
It's time to switch to xdg-shell v6 (in preparation for xdg-shell
stable, which is only available with Qt 5.12, or later).

However we have to bump the requirement from Qt 5.10 to Qt 5.11,
in order to use xdg-shell v6.

xdg-shell-v6 doesn't allow us to customize shell surfaces, so we have
to come up with a new system.  Regardless, this feels easier to
understand and maintain anyways.

We incapsulate the shell surface instead of inheriting from it.
This requires an overhaul of the window management.

The logic to determine the primary chrome is now written in C++,
in order to share it across implementations.

Animations were rewritten for Material Design.

Window menu now doesn't take away focus and some actions
are now implemented: such as maximize, restore and close.

Window switcher now shows the whole window, server-side decoration
include and handles the situation when the last selected
window is closed.

With a custom shell surface item we can now move windows by
pressing Alt+LMB reliably.

Closes: #146