Fix multiple screen regression
We can't assign a QScreen to Window::screen property because it's a
different type. We need to use the QtQuick counterpart.

The screen was not recognized and Qt assumed all windows were belonging
to the primary screen, but this is not allowed with eglfs which failed
with "EGLFS: OpenGL windows cannot be mixed with others."

Expose the screen index and use that to retrieve the QtQuick screen
from Qt.application.screens, this way we use the correct object
type.