Fix QPA plugin fallback
Wayland is tried first, but sometimes it fails with a protocol error
due to wl_drm not authenticating without a Wayland compositor running.

If XDG_SESSION_TYPE is set to tty we should be pretty sure we are
running from a vt, so in that case just go straight to "liri".