Private QML plugin
Move Wayland extensions to a static library and create a static QML
plugin that links to that and D-Bus interfaces.

The static plugin is then imported by the compositor.

This way we have separated the code but still build it inside the
compositor and do not allow external applications to use it.