Fix layer surface setup
Set anchors according to the protocol and calculate the surface size
based on the suggested size and anchors.

Configure the surface as soon as all layer surface attributes are
set.  This in turn will trigger the client to attach a buffer.

Also, do not paint surface item until the surface has been configured
in order to display only complete surfaces.