Make sure the the desktop is always revealed
If the shell helper client crashes we don't receive the ready event and
the splash screen keeps being visible, thus making the desktop unusable
to our users.

To avoid all of this, we set a timer that after a while reveals the
desktop regardless of what shell helper does.

The timer is set to 15 s for now, if shell helper works it should be
able to start in less than that.