Use the new layer shell QtWayland shell integration
We now have a proper QtWayland shell integration plugin for layer shell
clients.

This solution is better because it integrates very well with the
QtWayland platform plugin and hides all the details from the QtQuick
world.

Layer surfaces hidden and shown again are now properly recreated and all
their attributes are sent back to the compositor.

In the future it will also be easier to manage popup and modal surfaces.

Closes: #210