Redesign notifications
Draw elevation effect and close button on the client side.

This allows us to semplify quite a lot the decoration of this window on
the compositor side: we can now use a standard layer surface item.

At the same time we improve the close button look by using a RoundButton
instead of our own thing.