Client-side lock screen
Move all the lock screen code to the client.

We also design new animations for the lock screen, however the enter
animation is supposed to slide the window in the Y axis but it doesn't
work since it's anchored to the parent.

In the near future, the layer surfaces layout will be revamped to fix
this issue.

The indicators row cannot be done on the client right now, but there are
plans to bring it back once the panel is also client-side.

Issue: #184