Implement decorations with C++
The new input handlers do not work as expected.

Mouse click fall through: if you overlap the title bar of two windows and
click on the one stacked above, your click will hit the window below.

Windows can be dragged even when the modal dialog overlay is active.

Rewrite the decorations with C++ to avoid these issues.

Resize handlers seem to work just fine configuring the appropriate grab
permissions but just to make sure we disable them when the edge is
invisible or the shell surface item has input events disabled.

While we are here, we make sure input events on shell surface items are
disabled when the modal overlay is shown and replace the MouseArea with
input handlers that can take over everything.

Closes: #212