Improve splash screen
Use a wave effect for the splash screen and use the same color for the
window background to make the transition smooth.

Busy indicator animation is jerky because a lot is going on at startup
in the compositor.  Since the startup is so fast that it's not really
needed and we can get rid of it.

While we are here, we make sure the liri_shell.ready request is sent
after the interface is bound.

Closes: #213