Client-side panel
Rename Liri.Launcher QML module to Liri.Shell.TaskManager and move all
the code to the shell client.

Render the UI in a full screen layer surface, basically moving all QML
code from Shell.qml to the client process.

This makes it easier for us to manage popups, dialogs, menus, tooltips,
panel icon animations but we need to carefully specify the input region
and we cannot determine the available area computing the exclusive size.

A full screen layer surface may seem too much considering that most of
the time it renders a panel, but on mobile devices we will have more UI
elements, like a top panel or a drawer with application icons and this
is all easier to do it this way.

In the future, this single layer surface will be rendered in a hardware
layer without too much trouble.

Issue: #184