Refactoring
Clean up the code and remove components that we don't need.

Set the output window color so that it's initially black like the VT
color and provide a nice transition with the default screen color.
Then, change the color to match the splash screen once it is visible and
set the color to grey to match the default color as soon as the shell is
ready.