Synchronous lock screen image loading
Lock screen window was completely white, until the background image was
loaded.

To avoid this, we replace the Background component and provide ourselves
a simpler way to paint the background.

If a solid color is selected, then we use the actual window color,
otherwise a Rectangle with a gradient or an Image.

We don't track settings changes otherwise we would need a transition to
switch from an image to another, but that requires asynchronous loading.