Session locker
Use the new ext_session_lock_v1 protocol for session lockers,
replace our old in-house protocol with it.

The session locker is now a separate process, that is spawned as soon as
users will lock the session.

We also introduce HelperLauncher, to spawn helper process and restart
them if they crash.